Add spawnWeapon command & Fix Weapon's z hitbox

+++ nenequest/headers/weapon.hpp
@@ -13,7 +13,7 @@ class Weapon : public Entity {
         int getDamage();
         WeaponType getWeaponType();
 		void Animate(float lerp);
-		void setPosition(float x, float y);
+		void setPosition(int x, int y);
 
     private :
         WeaponType type;
